smtbx: Fix Python 3 syntax incompatible tests
Convert tests to pytest, including one that was not known to run_tests.py
