+++ smtbx/conftest.py
@@ -1,0 +1,4 @@
+from __future__ import absolute_import, division, print_function
+
+from libtbx.test_utils.pytest import libtbx_collector
+pytest_collect_file = libtbx_collector()
