+++ smtbx/refinement/constraints/tests/test_direction.py
@@ -1,13 +1,12 @@
-from __future__ import division
-from __future__ import print_function
-from cctbx.array_family import flex
-from cctbx import uctbx, xray, crystal
-from smtbx.refinement import constraints
-from scitbx.matrix import col
-from scitbx.matrix import row
-from libtbx.test_utils import approx_equal
+from __future__ import absolute_import, division, print_function
 
-def exercise_direction():
+def test_direction():
+  from cctbx.array_family import flex
+  from cctbx import uctbx, xray, crystal
+  from smtbx.refinement import constraints
+  from scitbx.matrix import col, row
+  from libtbx.test_utils import approx_equal
+
   uc = uctbx.unit_cell((1, 2, 3))
   xs = xray.structure(
     crystal_symmetry=crystal.symmetry(
@@ -45,11 +44,3 @@ def exercise_direction():
   nc = row(v01).cross(row(v21)).normalize()
   assert approx_equal(n, n1, eps=1e-15)
   assert approx_equal(n, nc, eps=1e-15)
-
-
-def run():
-  exercise_direction()
-  print 'OK'
-
-if __name__ == '__main__':
-  run()
