+++ smtbx/tests/test_utils.py
@@ -1,10 +1,10 @@
-from __future__ import division
-from __future__ import print_function
+from __future__ import absolute_import, division, print_function
+
 from cctbx import sgtbx
 from libtbx.test_utils import approx_equal
 from smtbx import development, utils
 
-def exercise_connectivity_table():
+def test_connectivity_table():
   xs = development.sucrose()
   connectivity = utils.connectivity_table(xs)
   pair_counts = [
@@ -21,10 +21,3 @@ def exercise_connectivity_table():
   connectivity.remove_bond(0, 1)
   connectivity.remove_bond(5,5, rt_mx_ji=sgtbx.rt_mx("x+1,y,z"))
   assert approx_equal(connectivity.pair_asu_table.pair_counts(), pair_counts)
-
-def run():
-  exercise_connectivity_table()
-  print "OK"
-
-if __name__ == '__main__':
-  run()
